extract db values test
